+++ octodns/processor/base.py
@@ -2,12 +2,9 @@
 #
 #
 
-from __future__ import (
-    absolute_import,
-    division,
-    print_function,
-    unicode_literals,
-)
+
+class ProcessorException(Exception):
+    pass
 
 
 class BaseProcessor(object):